% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caesar.R
\name{caesar_decrypt}
\alias{caesar_decrypt}
\title{Caesar Cipher Decryption}
\usage{
caesar_decrypt(alphabet, shift, encrypted_message)
}
\arguments{
\item{alphabet}{Character string containing the alphabet used for encryption}

\item{shift}{Integer specifying the shift value that was used for encryption}

\item{encrypted_message}{Character string to be decrypted}
}
\value{
Decrypted message
}
\description{
Decrypts a message encrypted with the Caesar cipher
}
\examples{
decrypted_msg <- caesar_decrypt("abcdefghijklmnopqrstuvwxyz", 3, "khoor")
print(decrypted_msg)
}
