% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.text_to_num}
\alias{.text_to_num}
\title{Convert text to numeric vector}
\usage{
.text_to_num(text)
}
\arguments{
\item{text}{input message}
}
\value{
numeric vector
}
\description{
Returns numeric vector
}
