% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill.R
\name{hill_decrypt}
\alias{hill_decrypt}
\title{Hill Cipher Decryption}
\usage{
hill_decrypt(ciphertext, beginning, key_matrix, mod)
}
\arguments{
\item{ciphertext}{encrypted message}

\item{key_matrix}{key matrix (inverse of encryption key matrix)}
}
\value{
decrypted message
}
\description{
Decrypts a ciphertext message that was encrypted using the Hill cipher algorithm.
}
