% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill.R
\name{hill_encrypt}
\alias{hill_encrypt}
\title{Hill Cipher Encryption}
\usage{
hill_encrypt(text, key_matrix)
}
\arguments{
\item{text}{input message}

\item{key_matrix}{key matrix}
}
\value{
encrypted message
}
\description{
Encrypts a plaintext message using the Hill cipher algorithm, a polygraphic substitution cipher based on linear algebra.
}
