% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill.R
\name{hill_decrypt}
\alias{hill_decrypt}
\title{Hill Cipher Decryption}
\usage{
hill_decrypt(ciphertext, beginning, key_matrix, mod)
}
\arguments{
\item{ciphertext}{the message that we will decode}

\item{beginning}{alphabet detection by initial letter}

\item{key_matrix}{square matrix of size NxN}

\item{mod}{the value of the module that we will be working with}
}
\value{
our decrypted message
}
\description{
Decrypts a ciphertext message that was encrypted using the Hill cipher algorithm.
}
\examples{
msg <- "PEALOJ"
key <- matrix(c(1,3,2,4), nrow=2)
decrypted_msg <- hill_decrypt(msg, "A", key, 33)
print(decrypted_msg)
}
