% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill.R
\name{hill_encrypt}
\alias{hill_encrypt}
\title{Hill Cipher Encryption}
\usage{
hill_encrypt(text, beginning, key_matrix, mod)
}
\arguments{
\item{text}{the message that we will encrypt}

\item{beginning}{alphabet detection by initial letter}

\item{key_matrix}{square matrix of size NxN}

\item{mod}{the value of the module that we will be working with}
}
\value{
our encrypted message
}
\description{
Encrypts a plaintext message using the Hill cipher algorithm, a polygraphic substitution cipher based on linear algebra.
}
\examples{
msg <- "HELLO"
key <- matrix(c(1,3,2,4), nrow=2)
encrypted_msg <- hill_encrypt(msg, "A", key, 33)
print(encrypted_msg)
}
